Lots...
Continued developing/fixing fire command.
Developed Auto to shoot from corner and go back to get 2.
Added logic to drive to closest position on swerve modules.
Probably lots more...

+++ TMW2013/Commands/SetWheelOffsets.h
@@ -30,10 +30,10 @@ public:
 	virtual bool IsFinished();
 	virtual void End();
 	virtual void Interrupted();
-	float FLOffset;
-	float FROffset;
-	float RLOffset;
-	float RROffset;
+	int FLOffset;
+	int FROffset;
+	int RLOffset;
+	int RROffset;
 	RAWCConstants* File;
 	bool complete;
